fix: correctly parse volume with decimal seperator
VLC has at some point started sending the volume as a number with a
decimal seperator, either a comma or period depending on the locale.
This causes a crash when reading the volume directly or from the status
command, because int() can't be called on a string that contains decimal
seperators.
Apparently, the volume can't be non-integer anyway: Setting it via the
GUI results in a fractional part of 0, and trying to set a non-integer
value via the telnet interface results in an error.

By replacing any commas with a period, the problem of users having to
adjust the locale of VLC is (hopefully) fixed.
Instead of casting the volume string into an int directly, it is now
first casted into a float and then into an int. This deals with the
crash.